+++ Application/Sources/Network/APIRouter.swift
@@ -11,9 +11,9 @@ import AppwiseCore
 
 enum APIRouter: AppwiseCore.Router {
 	static var baseURLString = env(
-		.dev("https://test.com/api"),
-		.stg("https://test.com/api"),
-		.prd("https://test.com/api")
+		.dev("https://test.development.appwi.se/api/"),
+		.stg("https://test.staging.appwi.se/api/"),
+		.prd("https://test.production.appwi.se/api/")
 	)
 
 	case user